Add files via upload
Change of the date format to use UNIX seconds formated timestamp instead, to prevent alteration of the file created (containing the time limit used for group membership provisioning). Some more comments in the code, to let it more readable for everyone. To run the script requires ExecutionPolicy without any restrcitions.